+++ python/cudf/cudf/tests/groupby/conftest.py
@@ -1,0 +1,7 @@
+# Copyright (c) 2025, NVIDIA CORPORATION.
+import pytest
+
+
+@pytest.fixture(params=[True, False])
+def as_index(request):
+    return request.param
