+++ python/cudf/cudf/tests/test_groupby.py
@@ -106,58 +106,6 @@ def pdf(gdf):
     return gdf.to_pandas()
 
 
-def test_groupby_agg_min_max_dictargs():
-    nelem = 20
-    expect_df = (
-        make_frame(pd.DataFrame, nelem=nelem, extra_vals="ab")
-        .groupby(["x", "y"])
-        .agg({"a": "min", "b": "max"})
-    )
-    got_df = (
-        make_frame(DataFrame, nelem=nelem, extra_vals="ab")
-        .groupby(["x", "y"])
-        .agg({"a": "min", "b": "max"})
-    )
-    assert_groupby_results_equal(expect_df, got_df)
-
-
-def test_groupby_agg_min_max_dictlist():
-    nelem = 20
-    expect_df = (
-        make_frame(pd.DataFrame, nelem=nelem, extra_vals="ab")
-        .groupby(["x", "y"])
-        .agg({"a": ["min", "max"], "b": ["min", "max"]})
-    )
-    got_df = (
-        make_frame(DataFrame, nelem=nelem, extra_vals="ab")
-        .groupby(["x", "y"])
-        .agg({"a": ["min", "max"], "b": ["min", "max"]})
-    )
-    assert_groupby_results_equal(got_df, expect_df)
-
-
-@pytest.mark.parametrize("as_index", [True, False])
-def test_groupby_as_index_single_agg(pdf, gdf, as_index):
-    gdf = gdf.groupby("y", as_index=as_index).agg({"x": "mean"})
-    pdf = pdf.groupby("y", as_index=as_index).agg({"x": "mean"})
-    assert_groupby_results_equal(pdf, gdf, as_index=as_index, by="y")
-
-
-@pytest.mark.parametrize("engine", ["cudf", "jit"])
-@pytest.mark.parametrize("as_index", [True, False])
-@pytest.mark.skipif(
-    PANDAS_VERSION < PANDAS_CURRENT_SUPPORTED_VERSION,
-    reason="Include groups missing on old versions of pandas",
-)
-def test_groupby_as_index_apply(pdf, gdf, as_index, engine):
-    gdf = gdf.groupby("y", as_index=as_index).apply(
-        lambda df: df["x"].mean(), engine=engine
-    )
-    kwargs = {"func": lambda df: df["x"].mean(), "include_groups": False}
-    pdf = pdf.groupby("y", as_index=as_index).apply(**kwargs)
-    assert_groupby_results_equal(pdf, gdf, as_index=as_index, by="y")
-
-
 @pytest.mark.parametrize("as_index", [True, False])
 def test_groupby_as_index_multiindex(pdf, gdf, as_index):
     pdf = pd.DataFrame(
@@ -178,89 +126,6 @@ def test_groupby_as_index_multiindex(pdf, gdf, as_index):
         # column names don't match - check just the values
         for gcol, pcol in zip(gdf, pdf, strict=True):
             assert_array_equal(gdf[gcol].to_numpy(), pdf[pcol].values)
-
-
-def test_groupby_default(pdf, gdf):
-    gdf = gdf.groupby("y").agg({"x": "mean"})
-    pdf = pdf.groupby("y").agg({"x": "mean"})
-    assert_groupby_results_equal(pdf, gdf)
-
-
-def test_group_keys_true(pdf, gdf):
-    gdf = gdf.groupby("y", group_keys=True).sum()
-    pdf = pdf.groupby("y", group_keys=True).sum()
-    assert_groupby_results_equal(pdf, gdf)
-
-
-@pytest.mark.parametrize("as_index", [True, False])
-def test_groupby_getitem_getattr(as_index):
-    pdf = pd.DataFrame({"x": [1, 3, 1], "y": [1, 2, 3], "z": [1, 4, 5]})
-    gdf = cudf.from_pandas(pdf)
-    assert_groupby_results_equal(
-        pdf.groupby("x", as_index=as_index)["y"].sum(),
-        gdf.groupby("x", as_index=as_index)["y"].sum(),
-        as_index=as_index,
-        by="x",
-    )
-    assert_groupby_results_equal(
-        pdf.groupby("x", as_index=as_index).y.sum(),
-        gdf.groupby("x", as_index=as_index).y.sum(),
-        as_index=as_index,
-        by="x",
-    )
-    assert_groupby_results_equal(
-        pdf.groupby("x", as_index=as_index)[["y"]].sum(),
-        gdf.groupby("x", as_index=as_index)[["y"]].sum(),
-        as_index=as_index,
-        by="x",
-    )
-    assert_groupby_results_equal(
-        pdf.groupby(["x", "y"], as_index=as_index).sum(),
-        gdf.groupby(["x", "y"], as_index=as_index).sum(),
-        as_index=as_index,
-        by=["x", "y"],
-    )
-
-
-def test_groupby_cats():
-    rng = np.random.default_rng(seed=0)
-    df = DataFrame(
-        {"cats": pd.Categorical(list("aabaacaab")), "vals": rng.random(9)}
-    )
-
-    cats = df["cats"].values_host
-    vals = df["vals"].to_numpy()
-
-    grouped = df.groupby(["cats"], as_index=False).mean()
-
-    got_vals = grouped["vals"]
-
-    got_cats = grouped["cats"]
-
-    for i in range(len(got_vals)):
-        expect = vals[cats == got_cats[i]].mean()
-        np.testing.assert_almost_equal(got_vals[i], expect)
-
-
-def test_groupby_iterate_groups():
-    rng = np.random.default_rng(seed=0)
-    nelem = 20
-    df = DataFrame(
-        {
-            "key1": rng.integers(0, 3, nelem),
-            "key2": rng.integers(0, 2, nelem),
-            "val1": rng.random(nelem),
-            "val2": rng.random(nelem),
-        }
-    )
-
-    def assert_values_equal(arr):
-        np.testing.assert_array_equal(arr[0], arr)
-
-    for name, grp in df.groupby(["key1", "key2"]):
-        pddf = grp.to_pandas()
-        for k in "key1,key2".split(","):
-            assert_values_equal(pddf[k].values)
 
 
 @pytest.mark.skipif(
